requirements.txt
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
